a Fuckton of changes:
	- move 3ds relates stuff to ctr::, including threading
	- NDEBUG on release builds
	- (broken) show_extmeta_lazy
	- fix "Want to try req again" confirm box not being centered vertically
	- remove some tid stringifing and made it use native u64 funcs
	- preserve cursor position in selection menus
	- isLuma check is done by init_services
	- remove help page, just a dump of unneeded info at this point
	- mark firstRun config as unused
	- fix global buttons sometimes not being hidden
	- add finalize() call to builder<> and ScopedWidget
	- remove empty setup as virtual in BaseWidget
	- default z is now ui::layer::middle
	- ui::RenderQueue::detach_after
	- ui::RenderQueue::back() now casts with a template
	- after_render_complete returns a bool now
	- ui::builder<T>: align_y, align_x, add_to overloads for saving the widget pointer
	- ui::next::List<T> (mostly?) done
	- fix some ugly stuff in show_hlink()
	- remove waiting for wifi menu, as index fetching now fails normally
	- rewrote show_more(), next::sel_*() for new ui framework
	- show_searchbar_search() now uses next::sel_gam() instead of a next::sel_gam() copy/paste
	- fix bug in Text::prepare_arrays(): not enough memory was being allocated
	- implement Button::set_border()

+++ source/error.cc
@@ -314,7 +314,7 @@ error_container get_error(Result res)
 
 std::string pad8code(Result code)
 {
-	char errpad[9]; snprintf(errpad, 9, "%08lX", code);
+	char errpad[9]; snprintf(errpad, 9, "%08lX", (unsigned long) code);
 	return errpad;
 }
